+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+Targets GOV.UK Frontend v5.2.0.
+
 ## 2.1.0
 
 #### Page template